\name{ldacv.gui}
\alias{ldacv.gui}
\title{Gui interface to calculate the cross validation of the linear discriminate analysis after a principle component analysis}
\usage{ldacv.gui(dummy=NULL)}
\arguments{
\item{dummy}{}
}
\description{This is a gui interface to calculate the cross validation of the linear discriminate analysis after a principle component analysis}
\seealso{}
\examples{}
\keyword{}	
