\name{ldacv.gui}
\alias{ldacv.gui}
\title{Gui interface to calculate the cross validation of the linear discriminant analysis after principal components analysis}
\usage{ ldacv.gui(dummy=NULL)}
\arguments{
\item{dummy}{a dummy variable to invoke gui interface}
}
\description{This is a gui interface to calculate the cross validation of the fisher discriminant analysis after principal components analysis}
\seealso{}
\examples{
## open the gui interface
## calculate CV and then	
##invoke gui interface
}
\keyword{}	
